+++ CHANGELOG.md
@@ -7,8 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+## v0.2.0 (2024-10-30)
+
 - Added Docker container image
-- Slimmed down the JDK used in the nix package
+- Updated datomic-pro to version 1.0.7260
+- Switched to Nix's JDK 21 headless package (now that it is sufficiently headless)
+
+
+### Added
+
+- Created this flake with datomic-pro version 1.0.7075
 
 ## v0.1.0 (2024-06-12)